pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nanocode"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "croniter==6.2.2",
    "openai==2.24.0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.3",
]

[project.scripts]
nanocode = "nanocode.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nanocode*"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I"]
